+++ El_Cautivo/El_Cautivo/Minigames/FreezingMG.cs
@@ -7,8 +7,6 @@ using Microsoft.Xna.Framework.Input;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace El_Cautivo.Minigames
 {
@@ -71,9 +69,9 @@ namespace El_Cautivo.Minigames
             ActTexture = new Vector2((1920 / 2), (1080 - ActionTexture.Height) / 2);
             OnOffPos = new Vector2(1920 - (OnTexture.Width * 2), 1080 - (OffTexture.Height * 2));
 
-            BackButton = new Button(BackPos / Game1.dScale, BackTexture, OnBackButton, 4 / Game1.dScale, Button.ButtonType.OnUp);
-            ActionButton = new Button(ActTexture / Game1.dScale, ActionTexture, OnActionButton, 2 / Game1.dScale, Button.ButtonType.OnUp);
-            OnOffButton = new Button(OnOffPos / Game1.dScale, OffTexture, OnOnOffButton, 2 / Game1.dScale, Button.ButtonType.OnUp);
+            BackButton = new Button(BackPos / Game1.dScale, BackTexture, OnBackButton, (int)(4 / Game1.dScale), Button.ButtonType.OnUp);
+            ActionButton = new Button(ActTexture / Game1.dScale, ActionTexture, OnActionButton, (int)(2 / Game1.dScale), Button.ButtonType.OnUp);
+            OnOffButton = new Button(OnOffPos / Game1.dScale, OffTexture, OnOnOffButton, (int)(2 / Game1.dScale), Button.ButtonType.OnUp);
         }
 
         private void OnBackButton()
