+++ El_Cautivo/El_Cautivo/Minigames/VaporizingMG.cs
@@ -7,11 +7,8 @@ using Microsoft.Xna.Framework.Input;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection.Metadata;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace El_Cautivo.Minigames 
+namespace El_Cautivo.Minigames
 {
     public  class VaporizingMG : IMiniGame
     {
@@ -77,9 +74,9 @@ namespace El_Cautivo.Minigames
             OnOffPos = new Vector2(1920 - (OnTexture.Width * 2), 1080 - (OffTexture.Height * 2));
 
             tempSlider = new Slider(SliderBasisTexture, SliderTexture, TSliderPos / Game1.dScale);
-            BackButton = new Button(BackPos / Game1.dScale, BackTexture, OnBackButton, 4 / Game1.dScale, Button.ButtonType.OnUp);
-            ActionButton = new Button(ActTexture / Game1.dScale, ActionTexture, OnActionButton, 2 / Game1.dScale, Button.ButtonType.OnUp);
-            OnOffButton = new Button(OnOffPos / Game1.dScale, OffTexture, OnOnOffButton, 2 / Game1.dScale, Button.ButtonType.OnUp);
+            BackButton = new Button(BackPos / Game1.dScale, BackTexture, OnBackButton, (int)(4 / Game1.dScale), Button.ButtonType.OnUp);
+            ActionButton = new Button(ActTexture / Game1.dScale, ActionTexture, OnActionButton, (int)(2 / Game1.dScale), Button.ButtonType.OnUp);
+            OnOffButton = new Button(OnOffPos / Game1.dScale, OffTexture, OnOnOffButton, (int)(2 / Game1.dScale), Button.ButtonType.OnUp);
         }
 
         private void OnBackButton()
